CLM-26572 identifiable dao refactoring (#9861)

+++ insight-brain-service/src/main/java/com/sonatype/insight/brain/integration/CIEvaluationStatService.java
@@ -49,7 +49,7 @@ class CIEvaluationStatService
         sinceUtcDate);
 
     int numAppsWithCI = policyEvaluationDAO.getCountOfApplicationsWithCITriggeredEvaluations(sinceUtcDate);
-    int numTotalApps = applicationDAO.getCount();
+    int numTotalApps = (int) applicationDAO.getCount();
     int numAppsWithoutCI = numTotalApps - numAppsWithCI;
 
     return new CIEvaluationStatDTO(numAppsWithoutCI, numTotalApps);
